requirements.txt
argparse==1.3.0
dnslib==0.9.3
dnspython3==1.12.0
ipaddress==1.0.7
ipwhois==0.9.1
python-whois==0.2
pythonwhois==2.4.3
